feat: implement file storage service for attachments
- Created StorageService interface with local filesystem implementation
- Integrated storage service with attachment handlers
- Added file serving endpoint with authorization checks
- Updated nginx config for file upload/download support
- Added storage volume mount and environment variables
- Implemented secure path handling and file sanitization
- Generated structured storage paths (tickets/{id}/{year}/{month}/{day})
- Fixed ticket service type conversion issue

The attachment system now properly stores files on disk with:
- Path traversal protection
- Filename sanitization
- Unique file naming with timestamps
- Proper content type detection
- Support for streaming large files

+++ internal/service/ticket_service.go
@@ -183,7 +183,7 @@ func (s *TicketService) createArticle(tx *sql.Tx, article *models.Article) error
 
 // GetTicket retrieves a ticket by ID
 func (s *TicketService) GetTicket(ticketID int) (*models.Ticket, error) {
-	return s.ticketRepo.GetByID(ticketID)
+	return s.ticketRepo.GetByID(uint(ticketID))
 }
 
 // ListTickets retrieves a paginated list of tickets
